+++ src/qibolab/_core/instruments/qrng/__init__.py
@@ -1,0 +1,7 @@
+from . import extractors, qrng
+from .extractors import *
+from .qrng import *
+
+__all__ = []
+__all__ += qrng.__all__
+__all__ += extractors.__all__
