+++ src/qibolab/instruments/qrng.py
@@ -1,7 +1,7 @@
 """Quantum random number generator drivers."""
 
-from qibolab._core.instruments import qrng
 from qibolab._core.instruments.qrng import *  # noqa: F403
+from qibolab._core.instruments.qrng import qrng
 
 __all__ = []
 __all__ += qrng.__all__
